# ARROW-5651: [Python] Fix Incorrect conversion from strided Numpy array