@Ktakuya332C 2019-07-29
@pitrou Thank you for your review!
Here are some benchmarks on Ubuntu virtual environment running on MacbookPro.
```
Current master
In [4]: a = np.arange(100000, dtype=np.float32)[1:-1:2]
In [5]: strided = a[-1:1:-1]
In [6]: %timeit b = pa.array(a)
The slowest run took 8663.68 times longer than the fastest. This could mean that an intermediate result is being cached.
1 loop, best of 3: 129 µs per loop
In [7]: %timeit b = pa.array(a, type=pa.float64())
The slowest run took 88.79 times longer than the fastest. This could mean that an intermediate result is being cached.
1000 loops, best of 3: 299 µs per loop
In [8]: %timeit b = pa.array(strided)
10000 loops, best of 3: 134 µs per loop
In [9]: %timeit b = pa.array(strided, type=pa.float64())
1000 loops, best of 3: 287 µs per loop

This PR
In [4]: a = np.arange(100000, dtype=np.float32)[1:-1:2]
In [5]: strided = a[-1:1:-1]
In [6]: %timeit b = pa.array(a)
The slowest run took 962.69 times longer than the fastest. This could mean that an intermediate result is being cached.
10000 loops, best of 3: 155 µs per loop
In [7]: %timeit b = pa.array(a, type=pa.float64())
The slowest run took 67.47 times longer than the fastest. This could mean that an intermediate result is being cached.
1000 loops, best of 3: 304 µs per loop
In [8]: %timeit b = pa.array(strided)
10000 loops, best of 3: 157 µs per loop
In [9]: %timeit b = pa.array(strided, type=pa.float64())
1000 loops, best of 3: 306 µs per loop
```
The warning `The slowest run took ... times longer than the fastest` may be caused by the environment (VirtualBox) running the benchmarks. To determine the actual causes, I plan to try the benchmarks on another environment in a few days. 

Above benchmarks show that this PR probably introduces performance regression. If the regression is real and unacceptable, I think it would be better to close this PR and try another. What would you think?