@pitrou 2019-07-29
> The warning The slowest run took ... times longer than the fastest may be caused by the unstableness? of the environment (VirtualBox) running the benchmarks.

Right, I think that's probably the explanation. I will try to run them here.