@pitrou 2019-07-29
Here are some benchmarks on Ubuntu 18.04. On master:
```python
>>> a = np.arange(100000, dtype=np.float32)                                                                                                                              
>>> strided = a[-1:1:-1]                                                                                                                                                 
>>> %timeit b = pa.array(a)                                                                                                                                              
9.07 µs ± 29 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
>>> %timeit b = pa.array(strided)                                                                                                                                        
52.5 µs ± 94.3 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
>>> %timeit b = pa.array(a, type=pa.float64())                                                                                                                           
35.2 µs ± 66 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
>>> %timeit b = pa.array(strided, type=pa.float64())                                                                                                                     
77 µs ± 101 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

On this PR:
```python
>>> a = np.arange(100000, dtype=np.float32)                                                                                                                              
>>> strided = a[-1:1:-1]                                                                                                                                                 
>>> %timeit b = pa.array(a)                                                                                                                                              
9.11 µs ± 47.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
>>> %timeit b = pa.array(strided)                                                                                                                                        
288 µs ± 203 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
>>> %timeit b = pa.array(a, type=pa.float64())                                                                                                                           
36 µs ± 64.9 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
>>> %timeit b = pa.array(strided, type=pa.float64())                                                                                                                     
312 µs ± 128 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

So there's a very large slowdown for strided inputs, even when no type conversion is requested. It would be nice if you could try to improve the situation.