@pitrou 2019-07-29
(side note: benchmarks should be run on a release build. Your numbers suggests that you may have run them on a debug build. I'm guilty of the same error from time to time :-))