@Ktakuya332C 2019-07-30
@pitrou Thank you for benchmarking my PR and for your kind advice. Because I assume a patch that can mitigate the deceleration would be very different from this patch, I close this PR for now.
If I can find a way to fix the main issue and the deceleration problem, then I will propose that with another PR.